feat: Add interactive CLI fuzzer dashboard and shared runners
